requirements.txt
langgraph
langchain
streamlit
faiss-cpu
sentence-transformers
ollama
numpy
python-dotenv
langsmith
ragas